Improve the user interface.
Display the current value of all inputs and outputs.

Improve the way we deal with long lines.
It was possible to make the input go one way
and the outputs go the other way.
Test case:
  Start fresh.
  Do not click anywhere.
  Move your mouse to a point just above -5.
  Move your mouse just below -5.
  Alternate back and forth and watch the two output jump around.

Change the way we draw the last part of the line.
Before you click, the last part of the line can move.
We now display that part as a dashed line.
After you click, the line turns solid.

Finish the mouse control.
Moving out of the grid will return the input to the last saved value.
Clicking the mouse will save the current point.
You can click and drag to continuously save every move.
This was always the intent.