Add a demo() function.
Currently this is aimed at a developer.
It's a prototype for something real.

Create a constant for each of the well known formulas.
It makes sense to ask for naturalLogFormula than 1 or "Natural Log".
A misspelled word would cause an error at compile time.
And if you reorder the list, it doesn't break any code.